Do not attempt mv on bin if it already exists, ignore cloud-nuke configs
When GitHub Actions runs the 'make cli-install' target it fails
because the GOBIN is already set to ./bin.  GitHub appears to
use a 'mv' that by default does not wish to clobber the file
that now exists (even with -f specified), so place a check in
front of the 'mv' to prevent it from ever running if the file
exists there.  This does not change the existing behavior of
the make target as it does not run if a file exists anyways.

Signed-off-by: Kyle Squizzato <ksquizzato@mirantis.com>