Place k0s.extensions images in a seperate image tarball
For the first iteration of airgap we're using k0s airgap
bundle for the management cluster but the ManagedClusters we
deploy will still need the extensions images that the extension
charts reference.

This patch modifies the bundle-images target so it creates two
bundles, one that will serve as the management cluster bundle
and one that will serve as the ManagedCluster bundle.

Signed-off-by: Kyle Squizzato <ksquizzato@mirantis.com>